Leaning/acceleration fixed
I've built an alternate version of the 'Lerp' function, which works in a far more intuitive way than the default.  It has been

+++ Project5.0/Assets/Scripts/MenuScriptFiles.cs
@@ -31,6 +31,8 @@ using System.IO;
  * #Application.persistentDataPath#/saves/savedgames/#saveslot#/#scene#/presentitems/#item#
  * 
  * Bugs:
+ * Due to the use of the 'Directory.Move()' function in the file-overriding functionality,
+ * saving and loading may be bugged on some systems.
 */
 public class MenuScriptFiles : MonoBehaviour
 {
@@ -134,19 +136,24 @@ public class MenuScriptFiles : MonoBehaviour
 
     public void LoadInfo()
     {
-        //if (Serialization.DirectoryExists(Application.persistentDataPath + "/saves/savedgames/auxiliary"))
-        //{
-        //    UnityEditor.FileUtil.DeleteFileOrDirectory(Application.persistentDataPath + "/saves/savedgames/auxiliary");
-        //}
-        Serialization.DeleteDirectory(Application.persistentDataPath + "/saves/savedgames/auxiliary");
+        if (Serialization.DirectoryExists(Application.persistentDataPath + "/saves/savedgames/auxiliary"))
+        {
+            //UnityEditor.FileUtil.DeleteFileOrDirectory(Application.persistentDataPath + "/saves/savedgames/auxiliary");
+            Serialization.DeleteDirectory(Application.persistentDataPath + "/saves/savedgames/auxiliary");
+        }
 
-        //UnityEditor.FileUtil.CopyFileOrDirectory(
-        //        Application.persistentDataPath + "/saves/savedgames/" + PlayerPrefs.GetString("saved_game_slot"),
-        //        Application.persistentDataPath + "/saves/savedgames/auxiliary");
-
-        Serialization.CopyDirectory(
-            Application.persistentDataPath + "/saves/savedgames/" + PlayerPrefs.GetString("saved_game_slot"),
-            Application.persistentDataPath + "/saves/savedgames/auxiliary");
+        try
+        {
+            Directory.Move(
+                Application.persistentDataPath + "/saves/savedgames/" + PlayerPrefs.GetString("saved_game_slot"),
+                Application.persistentDataPath + "/saves/savedgames/auxiliary");
+        }
+        catch(System.Exception e)
+        {
+            Serialization.CopyDirectory(
+                Application.persistentDataPath + "/saves/savedgames/" + PlayerPrefs.GetString("saved_game_slot"),
+                Application.persistentDataPath + "/saves/savedgames/auxiliary");
+        }
     }
 
     public void LoadMainMenu()
@@ -227,12 +234,18 @@ public class MenuScriptFiles : MonoBehaviour
             Serialization.DeleteDirectory(Application.persistentDataPath + "/saves/savedgames/" + slot);
         }
 
-        //UnityEditor.FileUtil.CopyFileOrDirectory(
-        //        Application.persistentDataPath + "/saves/savedgames/auxiliary",
-        //        Application.persistentDataPath + "/saves/savedgames/" + slot);
-        Serialization.CopyDirectory(
-            Application.persistentDataPath + "/saves/savedgames/auxiliary",
-            Application.persistentDataPath + "/saves/savedgames/" + slot);
+        try
+        {
+            Directory.Move(
+                Application.persistentDataPath + "/saves/savedgames/auxiliary",
+                Application.persistentDataPath + "/saves/savedgames/" + slot);
+        }
+        catch(System.Exception e)
+        {
+            Serialization.CopyDirectory(
+                Application.persistentDataPath + "/saves/savedgames/auxiliary",
+                Application.persistentDataPath + "/saves/savedgames/" + slot);
+        }
     }
 
     public void EstablishDate(Text text)
